+++ modules/juce_audio_formats/codecs/juce_QuickTimeAudioFormat.cpp
@@ -194,7 +194,7 @@ public:
 
             bufferList->mNumberBuffers = 1;
             bufferList->mBuffers[0].mNumberChannels = inputStreamDesc.mChannelsPerFrame;
-            bufferList->mBuffers[0].mDataByteSize =  jmax ((UInt32) 4096, (UInt32) (samplesPerFrame * inputStreamDesc.mBytesPerFrame) + 16);
+            bufferList->mBuffers[0].mDataByteSize =  jmax ((UInt32) 4096, (UInt32) (samplesPerFrame * (int) inputStreamDesc.mBytesPerFrame) + 16);
 
             dataBuffer.malloc (bufferList->mBuffers[0].mDataByteSize);
             bufferList->mBuffers[0].mData = dataBuffer;
@@ -262,10 +262,10 @@ public:
                 }
 
                 int framesToDo = jmin (numSamples, (int) (bufferList->mBuffers[0].mDataByteSize / inputStreamDesc.mBytesPerFrame));
-                bufferList->mBuffers[0].mDataByteSize = inputStreamDesc.mBytesPerFrame * framesToDo;
+                bufferList->mBuffers[0].mDataByteSize = inputStreamDesc.mBytesPerFrame * (UInt32) framesToDo;
 
                 UInt32 outFlags = 0;
-                UInt32 actualNumFrames = framesToDo;
+                UInt32 actualNumFrames = (UInt32) framesToDo;
                 OSStatus err = MovieAudioExtractionFillBuffer (extractor, &actualNumFrames, bufferList, &outFlags);
                 if (err != noErr)
                 {
@@ -274,7 +274,7 @@ public:
                 }
 
                 lastSampleRead = startSampleInFile + actualNumFrames;
-                const int samplesReceived = actualNumFrames;
+                const int samplesReceived = (int) actualNumFrames;
 
                 for (int j = numDestChannels; --j >= 0;)
                 {
@@ -298,7 +298,7 @@ public:
                 {
                     for (int j = numDestChannels; --j >= 0;)
                         if (destSamples[j] != nullptr)
-                            zeromem (destSamples[j] + startOffsetInDestBuffer, sizeof (int) * numSamples);
+                            zeromem (destSamples[j] + startOffsetInDestBuffer, sizeof (int) * (size_t) numSamples);
 
                     break;
                 }
